Move calling coveralls to outside container.
It seems to work much less well on non-Travis CI systems. It requires
.git data to be present unless you're filling in everything manually, so
it's more convenient to just do it outside. We're also seeing strange
issues where it doesn't work unless the variables are exported, which
makes it odd that it worked in Travis in the first place.

Update scripts to use real environment variables.